+++ client/lib/link.hpp
@@ -21,7 +21,7 @@ namespace autd {
 		class Link {
 		protected:
 			virtual std::vector<uint16_t>& accessGainData(Gain& gain, const int deviceId);
-			virtual uint32_t& accessSent(Modulation& mod) noexcept;
+			virtual size_t& accessSent(Modulation& mod) noexcept;
 		public:
 			virtual void Open(std::string location) = 0;
 			virtual void Close() = 0;
